Refactor to simplify dev xp
