+++ packages/twenty-server/src/engine/core-modules/billing/billing.module.ts
@@ -2,16 +2,17 @@ import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 
 import { BillingController } from 'src/engine/core-modules/billing/billing.controller';
-import { BillingService } from 'src/engine/core-modules/billing/billing.service';
-import { StripeModule } from 'src/engine/core-modules/billing/stripe/stripe.module';
-import { BillingSubscription } from 'src/engine/core-modules/billing/entities/billing-subscription.entity';
-import { BillingSubscriptionItem } from 'src/engine/core-modules/billing/entities/billing-subscription-item.entity';
-import { Workspace } from 'src/engine/core-modules/workspace/workspace.entity';
 import { BillingResolver } from 'src/engine/core-modules/billing/billing.resolver';
+import { BillingSubscriptionItem } from 'src/engine/core-modules/billing/entities/billing-subscription-item.entity';
+import { BillingSubscription } from 'src/engine/core-modules/billing/entities/billing-subscription.entity';
 import { BillingWorkspaceMemberListener } from 'src/engine/core-modules/billing/listeners/billing-workspace-member.listener';
-import { UserWorkspaceModule } from 'src/engine/core-modules/user-workspace/user-workspace.module';
+import { BillingPortalWorkspaceService } from 'src/engine/core-modules/billing/services/billing-portal.workspace-service';
+import { BillingSubscriptionService } from 'src/engine/core-modules/billing/services/billing-subscription.service';
+import { BillingWebhookService } from 'src/engine/core-modules/billing/services/billing-webhook.service';
+import { StripeModule } from 'src/engine/core-modules/billing/stripe/stripe.module';
 import { FeatureFlagEntity } from 'src/engine/core-modules/feature-flag/feature-flag.entity';
-import { BillingWorkspaceService } from 'src/engine/core-modules/billing/billing.workspace-service';
+import { UserWorkspaceModule } from 'src/engine/core-modules/user-workspace/user-workspace.module';
+import { Workspace } from 'src/engine/core-modules/workspace/workspace.entity';
 
 @Module({
   imports: [
@@ -29,11 +30,16 @@ import { BillingWorkspaceService } from 'src/engine/core-modules/billing/billing
   ],
   controllers: [BillingController],
   providers: [
-    BillingService,
-    BillingWorkspaceService,
+    BillingSubscriptionService,
+    BillingWebhookService,
+    BillingPortalWorkspaceService,
     BillingResolver,
     BillingWorkspaceMemberListener,
   ],
-  exports: [BillingService, BillingWorkspaceService],
+  exports: [
+    BillingSubscriptionService,
+    BillingPortalWorkspaceService,
+    BillingWebhookService,
+  ],
 })
 export class BillingModule {}
