+++ packages/twenty-server/src/engine/core-modules/billing/dto/product.input.ts
@@ -2,7 +2,7 @@ import { ArgsType, Field } from '@nestjs/graphql';
 
 import { IsNotEmpty, IsString } from 'class-validator';
 
-import { AvailableProduct } from 'src/engine/core-modules/billing/billing.workspace-service';
+import { AvailableProduct } from 'src/engine/core-modules/billing/interfaces/available-product.interface';
 
 @ArgsType()
 export class ProductInput {
