+++ packages/twenty-server/src/engine/core-modules/billing/interfaces/available-product.interface.ts
@@ -1,0 +1,3 @@
+export enum AvailableProduct {
+  BasePlan = 'base-plan',
+}
