+++ packages/twenty-server/src/engine/core-modules/billing/stripe/stripe.service.ts
@@ -2,9 +2,12 @@ import { Injectable, Logger } from '@nestjs/common';
 
 import Stripe from 'stripe';
 
-import { EnvironmentService } from 'src/engine/integrations/environment/environment.service';
-import { User } from 'src/engine/core-modules/user/user.entity';
+import { AvailableProduct } from 'src/engine/core-modules/billing/interfaces/available-product.interface';
+
+import { ProductPriceEntity } from 'src/engine/core-modules/billing/dto/product-price.entity';
 import { BillingSubscriptionItem } from 'src/engine/core-modules/billing/entities/billing-subscription-item.entity';
+import { User } from 'src/engine/core-modules/user/user.entity';
+import { EnvironmentService } from 'src/engine/integrations/environment/environment.service';
 
 @Injectable()
 export class StripeService {
@@ -30,10 +33,28 @@ export class StripeService {
     );
   }
 
-  async getProductPrices(stripeProductId: string) {
-    return this.stripe.prices.search({
+  async getStripePrices(product: AvailableProduct) {
+    const stripeProductId = this.getStripeProductId(product);
+
+    const prices = await this.stripe.prices.search({
       query: `product: '${stripeProductId}'`,
     });
+
+    return this.formatProductPrices(prices.data);
+  }
+
+  async getStripePrice(product: AvailableProduct, recurringInterval: string) {
+    const productPrices = await this.getStripePrices(product);
+
+    return productPrices.find(
+      (price) => price.recurringInterval === recurringInterval,
+    );
+  }
+
+  getStripeProductId(product: AvailableProduct) {
+    if (product === AvailableProduct.BasePlan) {
+      return this.environmentService.get('BILLING_STRIPE_BASE_PLAN_PRODUCT_ID');
+    }
   }
 
   async updateSubscriptionItem(stripeItemId: string, quantity: number) {
@@ -118,5 +139,30 @@ export class StripeService {
         quantity: stripeSubscriptionItem.quantity,
       },
     );
+  }
+
+  formatProductPrices(prices: Stripe.Price[]) {
+    const result: Record<string, ProductPriceEntity> = {};
+
+    prices.forEach((item) => {
+      const interval = item.recurring?.interval;
+
+      if (!interval || !item.unit_amount) {
+        return;
+      }
+      if (
+        !result[interval] ||
+        item.created > (result[interval]?.created || 0)
+      ) {
+        result[interval] = {
+          unitAmount: item.unit_amount,
+          recurringInterval: interval,
+          created: item.created,
+          stripePriceId: item.id,
+        };
+      }
+    });
+
+    return Object.values(result).sort((a, b) => a.unitAmount - b.unitAmount);
   }
 }
