+++ packages/twenty-server/src/engine/core-modules/onboarding/onboarding.resolver.ts
@@ -1,13 +1,13 @@
 import { UseGuards } from '@nestjs/common';
 import { Mutation, Resolver } from '@nestjs/graphql';
 
-import { JwtAuthGuard } from 'src/engine/guards/jwt.auth.guard';
 import { OnboardingStepSuccess } from 'src/engine/core-modules/onboarding/dtos/onboarding-step-success.dto';
-import { AuthUser } from 'src/engine/decorators/auth/auth-user.decorator';
-import { User } from 'src/engine/core-modules/user/user.entity';
-import { AuthWorkspace } from 'src/engine/decorators/auth/auth-workspace.decorator';
-import { Workspace } from 'src/engine/core-modules/workspace/workspace.entity';
 import { OnboardingService } from 'src/engine/core-modules/onboarding/onboarding.service';
+import { User } from 'src/engine/core-modules/user/user.entity';
+import { Workspace } from 'src/engine/core-modules/workspace/workspace.entity';
+import { AuthUser } from 'src/engine/decorators/auth/auth-user.decorator';
+import { AuthWorkspace } from 'src/engine/decorators/auth/auth-workspace.decorator';
+import { JwtAuthGuard } from 'src/engine/guards/jwt.auth.guard';
 
 @UseGuards(JwtAuthGuard)
 @Resolver()
@@ -19,10 +19,11 @@ export class OnboardingResolver {
     @AuthUser() user: User,
     @AuthWorkspace() workspace: Workspace,
   ): Promise<OnboardingStepSuccess> {
-    await this.onboardingService.removeSyncEmailOnboardingStep(
-      user.id,
-      workspace.id,
-    );
+    await this.onboardingService.toggleOnboardingConnectAccountCompletion({
+      userId: user.id,
+      workspaceId: workspace.id,
+      value: true,
+    });
 
     return { success: true };
   }
