+++ packages/twenty-server/src/engine/core-modules/onboarding/onboarding.service.ts
@@ -1,47 +1,43 @@
 import { Injectable } from '@nestjs/common';
 
-import { BillingWorkspaceService } from 'src/engine/core-modules/billing/billing.workspace-service';
 import { SubscriptionStatus } from 'src/engine/core-modules/billing/entities/billing-subscription.entity';
+import { BillingSubscriptionService } from 'src/engine/core-modules/billing/services/billing-subscription.service';
 import { OnboardingStatus } from 'src/engine/core-modules/onboarding/enums/onboarding-status.enum';
 import { UserVarsService } from 'src/engine/core-modules/user/user-vars/services/user-vars.service';
 import { User } from 'src/engine/core-modules/user/user.entity';
-import {
-  Workspace,
-  WorkspaceActivationStatus,
-} from 'src/engine/core-modules/workspace/workspace.entity';
+import { WorkspaceActivationStatus } from 'src/engine/core-modules/workspace/workspace.entity';
+import { EnvironmentService } from 'src/engine/integrations/environment/environment.service';
 import { isDefined } from 'src/utils/is-defined';
 
 export enum OnboardingStepKeys {
-  SYNC_EMAIL_ONBOARDING_STEP = 'SYNC_EMAIL_ONBOARDING_STEP',
-  INVITE_TEAM_ONBOARDING_STEP = 'INVITE_TEAM_ONBOARDING_STEP',
-  CREATE_PROFILE_ONBOARDING_STEP = 'CREATE_PROFILE_ONBOARDING_STEP',
+  ONBOARDING_CONNECT_ACCOUNT_COMPLETE = 'ONBOARDING_CONNECT_ACCOUNT_COMPLETE',
+  ONBOARDING_INVITE_TEAM_COMPLETE = 'ONBOARDING_INVITE_TEAM_COMPLETE',
+  ONBOARDING_CREATE_PROFILE_COMPLETE = 'ONBOARDING_CREATE_PROFILE_COMPLETE',
 }
 
 export type OnboardingKeyValueTypeMap = {
-  [OnboardingStepKeys.SYNC_EMAIL_ONBOARDING_STEP]: boolean;
-  [OnboardingStepKeys.INVITE_TEAM_ONBOARDING_STEP]: boolean;
-  [OnboardingStepKeys.CREATE_PROFILE_ONBOARDING_STEP]: boolean;
+  [OnboardingStepKeys.ONBOARDING_CONNECT_ACCOUNT_COMPLETE]: boolean;
+  [OnboardingStepKeys.ONBOARDING_INVITE_TEAM_COMPLETE]: boolean;
+  [OnboardingStepKeys.ONBOARDING_CREATE_PROFILE_COMPLETE]: boolean;
 };
 
 @Injectable()
 export class OnboardingService {
   constructor(
-    private readonly billingWorkspaceService: BillingWorkspaceService,
+    private readonly billingSubscriptionService: BillingSubscriptionService,
+    private readonly environmentService: EnvironmentService,
     private readonly userVarsService: UserVarsService<OnboardingKeyValueTypeMap>,
   ) {}
 
   private async isSubscriptionIncompleteOnboardingStatus(user: User) {
-    const isBillingEnabledForWorkspace =
-      await this.billingWorkspaceService.isBillingEnabledForWorkspace(
-        user.defaultWorkspaceId,
-      );
+    const isBillingEnabled = this.environmentService.get('IS_BILLING_ENABLED');
 
-    if (!isBillingEnabledForWorkspace) {
+    if (!isBillingEnabled) {
       return false;
     }
 
     const currentBillingSubscription =
-      await this.billingWorkspaceService.getCurrentBillingSubscription({
+      await this.billingSubscriptionService.getCurrentBillingSubscription({
         workspaceId: user.defaultWorkspaceId,
       });
 
@@ -58,29 +54,6 @@ export class OnboardingService {
     );
   }
 
-  private async isProfileCreationOnboardingStatus(user: User) {
-    return await this.userVarsService.get({
-      userId: user.id,
-      workspaceId: user.defaultWorkspaceId,
-      key: OnboardingStepKeys.CREATE_PROFILE_ONBOARDING_STEP,
-    });
-  }
-
-  private async isSyncEmailOnboardingStatus(user: User) {
-    return await this.userVarsService.get({
-      userId: user.id,
-      workspaceId: user.defaultWorkspaceId,
-      key: OnboardingStepKeys.SYNC_EMAIL_ONBOARDING_STEP,
-    });
-  }
-
-  private async isInviteTeamOnboardingStatus(workspace: Workspace) {
-    return await this.userVarsService.get({
-      workspaceId: workspace.id,
-      key: OnboardingStepKeys.INVITE_TEAM_ONBOARDING_STEP,
-    });
-  }
-
   async getOnboardingStatus(user: User) {
     if (await this.isSubscriptionIncompleteOnboardingStatus(user)) {
       return OnboardingStatus.PLAN_REQUIRED;
@@ -90,67 +63,82 @@ export class OnboardingService {
       return OnboardingStatus.WORKSPACE_ACTIVATION;
     }
 
-    if (await this.isProfileCreationOnboardingStatus(user)) {
+    const userVars = await this.userVarsService.getAll({
+      userId: user.id,
+      workspaceId: user.defaultWorkspaceId,
+    });
+
+    const isProfileCreationComplete =
+      userVars.get(OnboardingStepKeys.ONBOARDING_CREATE_PROFILE_COMPLETE) ===
+      true;
+
+    const isConnectAccountComplete =
+      userVars.get(OnboardingStepKeys.ONBOARDING_CONNECT_ACCOUNT_COMPLETE) ===
+      true;
+
+    const isInviteTeamComplete =
+      userVars.get(OnboardingStepKeys.ONBOARDING_INVITE_TEAM_COMPLETE) === true;
+
+    if (!isProfileCreationComplete) {
       return OnboardingStatus.PROFILE_CREATION;
     }
 
-    if (await this.isSyncEmailOnboardingStatus(user)) {
+    if (!isConnectAccountComplete) {
       return OnboardingStatus.SYNC_EMAIL;
     }
 
-    if (await this.isInviteTeamOnboardingStatus(user.defaultWorkspace)) {
+    if (!isInviteTeamComplete) {
       return OnboardingStatus.INVITE_TEAM;
     }
 
     return OnboardingStatus.COMPLETED;
   }
 
-  async addSyncEmailOnboardingStep(user: User) {
+  async toggleOnboardingConnectAccountCompletion({
+    userId,
+    workspaceId,
+    value,
+  }: {
+    userId: string;
+    workspaceId: string;
+    value: boolean;
+  }) {
     await this.userVarsService.set({
-      userId: user.id,
-      workspaceId: user.defaultWorkspaceId,
-      key: OnboardingStepKeys.SYNC_EMAIL_ONBOARDING_STEP,
-      value: true,
+      userId,
+      workspaceId: workspaceId,
+      key: OnboardingStepKeys.ONBOARDING_CONNECT_ACCOUNT_COMPLETE,
+      value,
     });
   }
 
-  async addInviteTeamOnboardingStep(user: User) {
+  async toggleOnboardingInviteTeamCompletion({
+    workspaceId,
+    value,
+  }: {
+    workspaceId: string;
+    value: boolean;
+  }) {
     await this.userVarsService.set({
-      workspaceId: user.defaultWorkspaceId,
-      key: OnboardingStepKeys.INVITE_TEAM_ONBOARDING_STEP,
-      value: true,
+      workspaceId,
+      key: OnboardingStepKeys.ONBOARDING_INVITE_TEAM_COMPLETE,
+      value,
     });
   }
 
-  async addCreateProfileOnboardingStep(user: User) {
+  async toggleOnboardingCreateProfileCompletion({
+    userId,
+    workspaceId,
+    value,
+  }: {
+    userId: string;
+    workspaceId: string;
+    value: boolean;
+  }) {
     await this.userVarsService.set({
-      userId: user.id,
-      workspaceId: user.defaultWorkspaceId,
-      key: OnboardingStepKeys.CREATE_PROFILE_ONBOARDING_STEP,
-      value: true,
-    });
-  }
-
-  async removeCreateProfileOnboardingStep(userId: string, workspaceId: string) {
-    await this.userVarsService.delete({
       userId,
       workspaceId,
-      key: OnboardingStepKeys.CREATE_PROFILE_ONBOARDING_STEP,
-    });
-  }
-
-  async removeInviteTeamOnboardingStep(workspaceId: string) {
-    await this.userVarsService.delete({
-      workspaceId,
-      key: OnboardingStepKeys.INVITE_TEAM_ONBOARDING_STEP,
-    });
-  }
-
-  async removeSyncEmailOnboardingStep(userId: string, workspaceId: string) {
-    await this.userVarsService.delete({
-      userId,
-      workspaceId,
-      key: OnboardingStepKeys.SYNC_EMAIL_ONBOARDING_STEP,
+      key: OnboardingStepKeys.ONBOARDING_CREATE_PROFILE_COMPLETE,
+      value,
     });
   }
 }
